+++ lib/braze_api/client.rb
@@ -9,7 +9,9 @@ require 'braze_api/endpoints/users/alias'
 require 'braze_api/endpoints/users/identify'
 require 'braze_api/endpoints/users/export'
 require 'braze_api/endpoints/users/delete'
-require 'braze_api/endpoints/subscription_groups/status'
+require 'braze_api/endpoints/subscription_groups/status/get'
+require 'braze_api/endpoints/subscription_groups/status/list'
+require 'braze_api/endpoints/subscription_groups/status/update'
 
 module BrazeAPI
   # The top-level class that handles configuration and connection to the Braze API.
@@ -21,7 +23,9 @@ module BrazeAPI
     include BrazeAPI::Endpoints::Users::Identify
     include BrazeAPI::Endpoints::Users::Export
     include BrazeAPI::Endpoints::Users::Delete
-    include BrazeAPI::Endpoints::SubscriptionGroups::Status
+    include BrazeAPI::Endpoints::SubscriptionGroups::Status::Get
+    include BrazeAPI::Endpoints::SubscriptionGroups::Status::List
+    include BrazeAPI::Endpoints::SubscriptionGroups::Status::Update
 
     def initialize(api_key:, braze_url:, app_id:)
       @api_key = api_key
@@ -32,6 +36,12 @@ module BrazeAPI
     # Returns a parsed response from a post request
     def post(endpoint, params: {})
       response = client.post(endpoint, params.to_json)
+      JSON.parse(response.body)
+    end
+
+    # Returns a parsed response from a get request
+    def get(endpoint, params: {})
+      response = client.get(endpoint, params)
       JSON.parse(response.body)
     end